requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
numpy==1.26.4
scikit-learn==1.8.0
xgboost==2.1.0
tensorflow-cpu==2.19.0
groq==0.11.0
joblib==1.4.2
python-dotenv==1.0.1
pydantic==2.9.0